+++ tapis-appsapi/src/main/java/edu/utexas/tacc/tapis/apps/api/resources/AppResource.java
@@ -1251,7 +1251,7 @@ public class AppResource
     List<String> errMessages = app1.checkAttributeRestrictions();
 
     // Now validate attributes that have special handling at API level.
-    // Currently, no additional checks.
+    // NO-OP: Currently, no additional checks.
 
     // If validation failed log error message and return response
     if (!errMessages.isEmpty())